Split off DSLParser
