+++ lib/ridgepole.rb
@@ -18,6 +18,8 @@ require 'ridgepole/default_limit'
 require 'ridgepole/delta'
 require 'ridgepole/diff'
 require 'ridgepole/dsl_parser'
+require 'ridgepole/dsl_parser/context'
+require 'ridgepole/dsl_parser/table_definition'
 require 'ridgepole/dumper'
 require 'ridgepole/execute_expander'
 require 'ridgepole/external_sql_executer'
